New docstring format support
Created `extractSlashDoubleStarDoc` function, (tried to make sure it kept most
of the same logic/flow as `extractLeadingTripleSlashDoc`)

By default the extension checks for `/** **/` style docstrings, if it doesn't
find any it falls back to checking `///` style.

Modified docSkeleton to insert the `/** **/` style of docstrings instead of `///`.

Added support for basic XML comment block (`comment.block.xml-documentation.cicode`)- this isn't as detailed as the
previous syntax highlighting (as in, it doesn’t try to match specific words,
instead it tries to highlight any basic XML - hopefully that will make modifying
the docstring format easier if needed in the future.)

To achieve this I did have to modify the `comment.line.cicode` section of the
grammar, so that it does not greedily highlight the XML docstrings:
```diff
-      "begin": "/\\*",
-      "end": "\\*/",
+      "begin": "/\\*(?!\\*)",
+      "end": "(?<!\\*)\\*\/",
       "name": "comment.block.cicode"
```

If you can think of a better way to do this please do share - I have not dealt
with TextMate grammars before :).

Likewise, I haven’t written much typescript, so any changes/suggestions are more
than welcome!